+++ src/integrators/dgvi/integrators_dgvi.jl
@@ -50,20 +50,20 @@ mutable struct ParametersDGVI{DT, TT, D, S, R, ET <: NamedTuple} <: Parameters{D
         new{DT,TT,D,length(r⁻),length(c),ET}(equs, Δt, b, c, m, a, r⁻, r⁺, zero(TT), zeros(DT,D), zeros(DT,D), zeros(DT,D), zeros(DT,D))
     end
 
-    function ParametersDGVI{DT,D}(equs::NamedTuple, Δt::TT, basis::Basis{TT}, quadrature::Quadrature{TT}) where {DT,TT,D}
+    function ParametersDGVI{DT,D}(equs::NamedTuple, Δt::TT, basis::Basis{TT}, quadrature::QuadratureRule{TT}) where {DT,TT,D}
         # compute coefficients
         m = zeros(TT, nnodes(quadrature), nbasis(basis))
         a = zeros(TT, nnodes(quadrature), nbasis(basis))
         r⁻= zeros(TT, nbasis(basis))
         r⁺= zeros(TT, nbasis(basis))
 
-        for i in 1:nbasis(basis)
-            for j in 1:nnodes(quadrature)
-                m[j,i] = evaluate(basis, i, nodes(quadrature)[j])
-                a[j,i] = derivative(basis, i, nodes(quadrature)[j])
+        for i in eachindex(basis)
+            for j in eachindex(quadrature)
+                m[j,i] = basis[nodes(quadrature)[j], i]
+                a[j,i] = basis'[nodes(quadrature)[j], i]
             end
-            r⁻[i] = evaluate(basis, i, one(TT))
-            r⁺[i] = evaluate(basis, i, zero(TT))
+            r⁻[i] = basis[one(TT), i]
+            r⁺[i] = basis[zero(TT), i]
         end
 
         if get_config(:verbosity) > 1
@@ -382,20 +382,20 @@ struct IntegratorDGVI{DT, TT, D, S, R,
                 ST <: NonlinearSolver{DT},
                 IT <: InitialGuessODE{TT}} <: IODEIntegrator{DT,TT}
     basis::BT
-    quadrature::Quadrature{TT,R}
+    quadrature::QuadratureRule{TT,R}
 
     params::PT
     solver::ST
     iguess::IT
     caches::CacheDict{PT}
 
-    function IntegratorDGVI(basis::BT, quadrature::Quadrature{TT,R}, params::ParametersDGVI{DT,TT,D,S},
+    function IntegratorDGVI(basis::BT, quadrature::QuadratureRule{TT,R}, params::ParametersDGVI{DT,TT,D,S},
                     solver::ST, iguess::IT, caches) where {DT,TT,D,S,R,BT,ST,IT}
         new{DT, TT, D, S, R, BT, typeof(params), ST, IT}(basis, quadrature, params, solver, iguess, caches)
     end
 
-    function IntegratorDGVI{DT,D}(equations::NamedTuple, basis::Basis{TT,P}, quadrature::Quadrature{TT,R}, Δt::TT;
-                                  interpolation=HermiteInterpolation{DT}) where {DT,TT,D,P,R}
+    function IntegratorDGVI{DT,D}(equations::NamedTuple, basis::Basis{TT}, quadrature::QuadratureRule{TT,R}, Δt::TT;
+                                  interpolation=HermiteInterpolation{DT}) where {DT,TT,D,R}
 
         # get number of stages
         S = nbasis(basis)
@@ -416,7 +416,7 @@ struct IntegratorDGVI{DT, TT, D, S, R,
         IntegratorDGVI(basis, quadrature, params, solver, iguess, caches)
     end
 
-    function IntegratorDGVI(equation::IODE{DT,TT}, basis::Basis{TT}, quadrature::Quadrature{TT}, Δt::TT; kwargs...) where {DT,TT}
+    function IntegratorDGVI(equation::IODE{DT,TT}, basis::Basis{TT}, quadrature::QuadratureRule{TT}, Δt::TT; kwargs...) where {DT,TT}
         IntegratorDGVI{DT, ndims(equation)}(get_function_tuple(equation), basis, quadrature, Δt; kwargs...)
     end
 end
@@ -444,12 +444,12 @@ end
 
 function initial_guess!(int::IntegratorDGVI{DT,TT,D,S,R}, sol::AtomicSolutionPODE{DT,TT},
                         cache::IntegratorCacheDGVI{DT}=int.caches[DT]) where {DT,TT,D,S,R}
-    if nnodes(int.basis) > 0
-        for i in 1:S
+    if nbasis(int.basis) > 0
+        for i in eachindex(int.basis)
             evaluate!(int.iguess, sol.q̅, sol.v̅,
                                   sol.q, sol.v,
                                   cache.q̃,
-                                  nodes(int.basis)[i])
+                                  grid(int.basis)[i])
 
             for k in 1:D
                 int.solver.x[D*(i-1)+k] = cache.q̃[k]